+++ rbtree.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2010-2015 Richard Braun.
+ * Copyright (c) 2010-2017 Richard Braun.
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -29,6 +29,7 @@
 
 #include <assert.h>
 #include <stddef.h>
+#include <stdint.h>
 
 #include "macros.h"
 #include "rbtree.h"
@@ -92,7 +93,7 @@ rbtree_node_set_parent(struct rbtree_node *node, struct rbtree_node *parent)
     assert(rbtree_node_check_alignment(node));
     assert(rbtree_node_check_alignment(parent));
 
-    node->parent = (unsigned long)parent | (node->parent & RBTREE_COLOR_MASK);
+    node->parent = (uintptr_t)parent | (node->parent & RBTREE_COLOR_MASK);
 }
 
 /*
@@ -192,7 +193,7 @@ rbtree_insert_rebalance(struct rbtree *tree, struct rbtree_node *parent,
     assert(rbtree_node_check_alignment(parent));
     assert(rbtree_node_check_alignment(node));
 
-    node->parent = (unsigned long)parent | RBTREE_COLOR_RED;
+    node->parent = (uintptr_t)parent | RBTREE_COLOR_RED;
     node->children[RBTREE_LEFT] = NULL;
     node->children[RBTREE_RIGHT] = NULL;
